refactor(inference): extract Metal resident-decode engine into metal_resident
Move the Metal GPU resident-decode subsystem out of inference.rs into a new
src/inference/metal_resident.rs: the resident decode-forward method
(try_resident_decode_forward_metal), the resident-prefill body (now
try_metal_resident_prefill), and the resident_weight_bytes helper, plus a
ResidentDecodeState type alias so the session field no longer names metal::
directly. Also seam the remaining wire_mode_active call and a previously-missed
metal_retained block-dot offload through metal_seam, and cfg(macos)-gate the
now-macOS-only `use crate::metal` import.

Verbatim relocation; behaviour byte-identical on all targets (the moved code
compiles everywhere via metal.rs's non-macOS stubs). The shared/x86/CUDA path
in inference.rs is now Metal-free; the only metal:: refs left are two
cfg(target_os="macos") f32-linear fallbacks.

Verified on Windows: cargo check + clippy clean; cargo test --lib 494 passed /
0 failed (ran the suite this time); cargo fmt clean; no EOL churn.
inference.rs -291 lines.
